requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
customtkinter==5.2.2
darkdetect==0.8.0
deep-translator==1.11.4
deepl==1.23.0
easyocr==1.7.2
filelock==3.19.1
fsspec==2025.9.0
idna==3.11
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
mss==10.1.0
networkx==3.2.1
ninja==1.13.0
numpy==2.0.2
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pyclipper==1.3.0.post6
python-bidi==0.6.6
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
soupsieve==2.8
sympy==1.14.0
tifffile==2024.8.30
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
urllib3==2.5.0
